Separated the google maps network calls
