+++ app/src/main/java/com/bps/gotwinter2021/ui/maps/MapsViewModel.kt
@@ -9,11 +9,12 @@ import com.bps.gotwinter2021.common.secret.API.API_KEY
 import com.bps.gotwinter2021.data.model.GOTTheatre
 import com.bps.gotwinter2021.data.network.networkmodel.ServiceResult
 import com.bps.gotwinter2021.data.network.repo.GOTRepo
+import com.bps.gotwinter2021.data.network.repo.GoogleMapsRepo
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import timber.log.Timber
 
-class MapsViewModel(application: Application, val repo: GOTRepo)
+class MapsViewModel(application: Application, val repo: GoogleMapsRepo)
     : AndroidViewModel(application){
 
     private val _theaterFeed = MutableLiveData<GOTTheatre?>()
